+++ app/src/main/java/com/fibelatti/photowidget/model/AppShortcutInfo.kt
@@ -1,0 +1,11 @@
+package com.fibelatti.photowidget.model
+
+import android.graphics.drawable.Drawable
+import androidx.compose.runtime.Stable
+
+@Stable
+data class AppShortcutInfo(
+    val id: String,
+    val label: String,
+    val icon: Drawable?,
+)