+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* Ability to select app shortcuts when using the "Open an app" tap action
+
 ## [v1.42.4] - 2026-06-01
 [v1.42.4]: https://github.com/fibelatti/photo-widget/releases/tag/v1.42.4